// frontend/src/generated/models/createInactiveUserAccountRequest.ts
/**
 * Generated by orval v7.4.1 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export interface CreateInactiveUserAccountRequest {
  username: string;
  email: string;
  language?: string;
  timezone?: string;
  roles?: string[];
  permissions?: string[];
  tenantId?: string;
}